Phase7/pyproject.toml
[project]
name = "ai-restaurant-recommendation-phase7"
version = "0.1.0"
description = "AI Restaurant Recommendation Service (Phase 7 UI layer)"
requires-python = ">=3.10"
dependencies = [
    "streamlit",
    "requests",
    "python-dotenv",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
